Implement dump/import cmd design & add test mode
